mpibenchmark:   -- use long long for number of iterations
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@26363 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ mpibenchmark/mpi_bench_generic.c
@@ -33,7 +33,7 @@ struct mpi_bench_common_s mpi_bench_common =
 
 /* ** Iterations ******************************************* */
 
-static inline size_t _iterations(int iterations, size_t len)
+static inline long long _iterations(long long iterations, size_t len)
 {
   const uint64_t max_data = 512 * 1024 * 1024;
   if(len <= 0)
